% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaunchApp.R
\name{apps}
\alias{apps}
\title{Launch some Shiny apps}
\usage{
apps(name = "app1", ...)
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to \code{shiny::runApp()} or to \code{rmarkdown::run()}}
}
\value{
Shiny app
}
\description{
Launch some Shiny apps
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{apps()
apps("app1")
apps("app2")
}\if{html}{\out{</div>}}
}

