% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomeApps-package.R
\docType{package}
\name{SomeApps-package}
\alias{SomeApps}
\alias{SomeApps-package}
\title{SomeApps: Two Shiny apps}
\description{
Show how to distribute Shiny applications (standard or Rmd) within an R package.
}
\author{
\strong{Maintainer}: Anouar El Ghouch \email{anouar.elghouch@uclouvain.be}

}
\keyword{internal}
